requirements.txt
botbuilder-core
asyncio
aiohttp~=3.8.1
requests~=2.27.1
python-dotenv~=0.19.2